Generate abi output only for soname paths
Because we want to use abigail to detect ABI breaks, make sure to
generate .abi files for only library soname paths rather than the
version paths.

For example, we want the abi report for /usr/lib64/libzstd.so.1, not
/usr/lib64/libzstd.so.1.Y.Z; when the package bumps Y or Z, there should
be no ABI break. If there is, abigail will detect it.

In the future, when zstd bumps to libzstd.so.2, there is an obvious ABI
break, and abigail can help determine which packages need rebuilding,
patching, etc.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>