Move the creation of the abi subpackage from autospec to rpm
Since the files are created by an RPM macro run post-%install, it stands
to reason that they should be packaged by an RPM macro too. This way, we
even restore the ability to "make build" old .spec files that did not
have an explicit %package abi.

On the other hand, packages with explicit %package abi will need to be
re-autospec'ed. Otherwise, you'll see during the build:

 Start: rpmbuild -bs
 error: Package already exists: %package abi

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>