mvn.prov: refactor to support artifact classifiers
Some of the current mvnbin packages ship JAR files with classifiers in
their names, so the mvn provides info should incorporate that.

Because the classifier is part of the file name, not a path component,
the script was refactored to construct individual pieces of provides
string first, then subtracting out some of those pieces to determine
whether a classifier is present, and finally printing the full provides
string.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>